{"schema":"decodex/commit/1","summary":"Modularize lifecycle guardrail and query boundaries","authority":"manual"}
