+++ apps/decodex/src/orchestrator/execution_lifecycle/identity.rs
@@ -1,0 +1,14 @@
+use crate::orchestrator::{IssueRunPlan, ServiceConfig, records};
+
+pub(crate) fn lifecycle_event_identity<'a>(
+	project: &'a ServiceConfig,
+	issue_run: &'a IssueRunPlan,
+) -> records::LinearExecutionEventIdentity<'a> {
+	records::LinearExecutionEventIdentity {
+		service_id: project.service_id(),
+		issue_id: &issue_run.issue.id,
+		issue_identifier: &issue_run.issue.identifier,
+		run_id: &issue_run.run_id,
+		attempt_number: issue_run.attempt_number,
+	}
+}
